Updated kernel installation phase
wget commands have been removed. The script now assumes kernel configs,
initramfs or tar packages are provided.

If a kernel config is provided, the kernel build will use it. Otherwise, it will
use a default configuration.

An initramfs is now generated if the kernel is being built. That way, the
initramfs corresponds to the kernel version.

Signed-off-by: Byron Sanchez <byron@hackbytes.com>